Extract inner classes and organise by package
