+++ fhirpath/src/test/java/au/csiro/pathling/test/yaml/annotations/YamlSpec.java
@@ -1,10 +1,10 @@
-package au.csiro.pathling.test.yaml;
+package au.csiro.pathling.test.yaml.annotations;
 
-import au.csiro.pathling.test.yaml.YamlSpecTestBase.FhirpathArgumentProvider;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ArgumentsSource;
+import au.csiro.pathling.test.yaml.FhirpathArgumentProvider;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ArgumentsSource;
 
 @Retention(RetentionPolicy.RUNTIME)
 @ArgumentsSource(FhirpathArgumentProvider.class)
