+++ fhirpath/src/test/java/au/csiro/pathling/test/yaml/resolver/ResolverBuilder.java
@@ -1,0 +1,17 @@
+package au.csiro.pathling.test.yaml.resolver;
+
+import au.csiro.pathling.fhirpath.context.ResourceResolver;
+import au.csiro.pathling.test.yaml.RuntimeContext;
+import jakarta.annotation.Nonnull;
+import java.util.function.Function;
+
+/**
+ * Interface for building resource resolvers with specific context.
+ */
+@FunctionalInterface
+public interface ResolverBuilder {
+
+  @Nonnull
+  ResourceResolver create(
+      @Nonnull final Function<RuntimeContext, ResourceResolver> resolveFactory);
+}
